requirements.txt
tensorflow==2.4.0
numpy==1.19.2
connected-components-3d
tensorflow-addons
matplotlib
protobuf
regex
pandas
tqdm
opencv-python
git+https://github.com/yoshihikoueno/dsargparse.git@master
p_tqdm
pyyaml
tensorboard_plugin_profile
ruamel.yaml
scipy
scikit-learn